+++ s3pathlib/core/base.py
@@ -2285,96 +2285,6 @@ class S3Path:
                 kwargs["compression"] = compression
         return smart_open.open(**kwargs)
 
-    def read_text(
-        self,
-        encoding="utf-8",
-        errors=None,
-        bsm: Optional['BotoSesManager'] = None,
-    ) -> str:
-        with self.open(
-            mode="r",
-            encoding=encoding,
-            errors=errors,
-            bsm=bsm,
-        ) as f:
-            return f.read()
-
-    def read_bytes(self, bsm: Optional['BotoSesManager'] = None) -> bytes:
-        with self.open(mode="rb", bsm=bsm) as f:
-            return f.read()
-
-    def write_text(
-        self,
-        data: str,
-        encoding="utf-8",
-        errors=None,
-        metadata: Optional[MetadataType] = None,
-        tags: Optional[TagType] = None,
-        bsm: Optional['BotoSesManager'] = None,
-    ):
-        """
-        Write text to s3 object. A simple wrapper around
-        `s3_client.put_object <https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/s3.html#S3.Bucket.put_object>`_
-
-        :param data: the text you want to write.
-        :param encoding: how do you want to encode text?
-        :param errors: how do you want to handle encode error? can be 'strict',
-            ``ignore``, ``replace``, ``xmlcharrefreplace``, ``backslashreplace``.
-            see more details `here <https://docs.python.org/3/library/stdtypes.html#str.encode>`_.
-        :param metadata: the s3 object metadata in string key value pair dict.
-        :param tags: the s3 object tags in string key value pair dict.
-
-        .. versionadded:: 1.0.3
-
-        .. versionchanged:: 1.1.1
-        """
-        s3_client = resolve_s3_client(context, bsm)
-        if errors:
-            body = data.encode(encoding, errors=errors)
-        else:
-            body = data.encode(encoding)
-        response = better_client.put_object(
-            s3_client=s3_client,
-            bucket=self.bucket,
-            key=self.key,
-            body=body,
-            metadata=metadata,
-            tags=tags,
-        )
-        self._tags = tags
-        return response
-
-    def write_bytes(
-        self,
-        data: bytes,
-        metadata: Optional[dict] = None,
-        tags: Optional[TagType] = None,
-        bsm: Optional['BotoSesManager'] = None,
-    ):
-        """
-        Write binary data to s3 object. A simple wrapper around
-        `s3_client.put_object <https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/s3.html#S3.Bucket.put_object>`_
-
-        :param data: the text you want to write.
-        :param metadata: the s3 object metadata in string key value pair dict.
-        :param tags: the s3 object tags in string key value pair dict.
-
-        .. versionadded:: 1.0.3
-
-        .. versionchanged:: 1.1.1
-        """
-        s3_client = resolve_s3_client(context, bsm)
-        response = better_client.put_object(
-            s3_client=s3_client,
-            bucket=self.bucket,
-            key=self.key,
-            body=data,
-            metadata=metadata,
-            tags=tags,
-        )
-        self._tags = tags
-        return response
-
     def touch(
         self,
         exist_ok: bool = True,
